refactor(organization-main): feature 하위 폴더 이름을 dot notation으로 변경
