+++ src/feature/organization/organization.main/components/icons/EmptyIcons.tsx
