+++ src/pages/organizations/[organizationId]/index.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from "next/router";
-import ViewOrganizationMainPage from "feature/organization/organizationMain/view/ViewOrganizationMainPage";
+import ViewOrganizationMainPage from "feature/organization/organization.main/view/ViewOrganizationMainPage";
 
 const PageOrganizationDetail = () => {
   const { query } = useRouter();
